Move SQLConf back. Assign default sizeInBytes to SparkLogicalPlan.
